% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-fns.R
\name{extract_palette}
\alias{extract_palette}
\title{Extracts palette from image}
\usage{
extract_palette(image_loc, n_cols)
}
\arguments{
\item{image_loc}{Location of image on disk}

\item{n_cols}{Number of colours to extract}
}
\value{
Returns a character vector of hex codes
}
\description{
The image is read in using \code{magick}, converted to RGB and clustered using kmeans. The user
must specify the number of clusters. The cluster centroids become the palette values.
}
\examples{
\dontrun{}
}
