% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-fns.R
\name{generate_palette_name}
\alias{generate_palette_name}
\title{Generate palette name}
\usage{
generate_palette_name(n)
}
\arguments{
\item{n}{Number of names to generate.}
}
\description{
Generates a palette name for evolved palettes. Names are generated from a random combination of adjectives and food related words.
}
\details{
This function randomly draws an adjective and a food related word from a dictionary The name is assigned to a newly generated
palette. Palette name can be changed when saving.
}
\examples{
\dontrun{generate_palette_name(5)}
}
