% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-fns.R
\name{get_pal_order}
\alias{get_pal_order}
\title{Palette colour order}
\usage{
get_pal_order(pal, rgb_hsv = "both")
}
\arguments{
\item{pal}{Palette. Character vector of hex codes.}

\item{rgb_hsv}{Sort by either rgb, hsv or both}
}
\description{
Sorts colour palette by hue.
}
\examples{
\dontrun{
pal <- c("#564862", "#EEFBFD", "#594543", "#8E4B3E", "#BF6856")
list(
    unsorted = pal,
    sorted = get_pal_order(pal)
    ) \%>\%
    imap(~show_palette(.x, .y)) \%>\%
    wrap_plots()
}
}
