% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-fns.R
\name{get_palette_data}
\alias{get_palette_data}
\title{Read palette data}
\usage{
get_palette_data()
}
\description{
Pulls together evoPalette and paletteer palettes ready to sample
}
\examples{
\dontrun{get_palette_data()}
}
