% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-fns.R
\name{open_palette_box}
\alias{open_palette_box}
\title{Palette box}
\usage{
open_palette_box(clear = FALSE)
}
\arguments{
\item{clear}{Set to TRUE to clear the palette box. Default FALSE}
}
\description{
Opens the palette box containing saved palettes from the evoPalette Shiny app
}
\details{
Any palettes saved while using the shiny app are saved here. Opening the box displays a list of the saved palettes. More
Palettes will be saved here everytime you use the app during a single session. Remember to save the palette box before closing the
session.
}
