% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-fns.R
\name{palette_box}
\alias{palette_box}
\title{Palette box}
\usage{
palette_box(clear = FALSE)
}
\arguments{
\item{clear}{Set to TRUE to clear the palette box. Default FALSE}
}
\value{

}
\description{
Opens the palette box containing saved palettes from the evoPalette Shiny app
}
\details{
Any palettes saved while using the shiny app are saved here. Opening the box displays a list of the saved palettes. More
Palettes will be saved here everytime the app is used with a session. Remember to save the palette box to disk before closing the
session to permanently save the palette.
}
\examples{
\dontrun{}
}
