% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-fns.R
\docType{data}
\name{palette_data}
\alias{palette_data}
\title{Read palette data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1453 rows and 5 columns.}
\usage{
palette_data
}
\description{
Pulls together evoPalette and paletteer palettes ready for sampling
}
\keyword{datasets}
