% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-fns.R
\name{random_palette}
\alias{random_palette}
\title{Random palette}
\usage{
random_palette(n_cols, n_palettes, feeling_lucky = FALSE)
}
\arguments{
\item{n_cols}{Number of colours in the palette. Numeric.}

\item{n_palettes}{Number of palettes to generate. Numeric.}

\item{feeling_lucky}{If TRUE generates completely random colours for the palette. Default FALSE}
}
\description{
Selects a random palette either from an exisiting set of palettes or completely random.
}
\details{
Palettes can be randomly selected from existing palettes or completely random colours if you're
feeling very lucky.
}
\examples{
\dontrun{
pals <- random_palette(5, 4)
map(pals, ~show_palette(.x)) \%>\%
    wrap_plots
}
}
