% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{evo_pal}
\alias{evo_pal}
\alias{scale_fill_evo}
\alias{scale_colour_evo}
\title{evoPalette palette function}
\usage{
evo_pal(name, discrete = TRUE, reverse = FALSE, ...)

scale_fill_evo(name, discrete = TRUE, reverse = FALSE, ...)

scale_colour_evo(name, discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{name}{Name of palette in the palette box}

\item{discrete}{Logical. Is the aesthetic discrete?}

\item{reverse}{Logical. Reverse the palette?}

\item{...}{Dots}
}
\description{
Generates a colour palette function for a discrete or continuous scale

Generalises the scale aesthetics for evoPalettes

Generalises the scale aesthetics
}
