% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-fns.R
\name{sort_palette}
\alias{sort_palette}
\title{Palette colour order}
\usage{
sort_palette(pal, rgb_hsv = "choose")
}
\arguments{
\item{pal}{Palette. Character vector of hex codes.}

\item{rgb_hsv}{Sort by either rgb, hsv, both, none or normal. See details.}
}
\value{

}
\description{
Sorts colour palette by hue.
}
\details{
Sorting colours is near impossible to get it right everytime. This allows sorting by a few methods. Default is \code{both}.
\itemize{
\item{\code{rgb}}: Sorts by RGB by first finding the lightest, then it's nearest neighbour and so on.
\item{\code{hsv}}: Sorts by HSV by first finding the most saturated, then it's nearest neighbour and so on.
\item{\code{both}}: Both RGB and HSV.
\item{\code{normal}}: Simple character sort on hex codes.
\item{\code{none}}: Returns the same order. Can be convenient.
}
}
\examples{
\dontrun{
pal <- c("#564862", "#EEFBFD", "#594543", "#8E4B3E", "#BF6856")
list(
    unsorted = pal,
    sorted = sort_palette(pal)
    ) \%>\%
    imap(~show_palette(.x, .y)) \%>\%
    wrap_plots()
}
}
