% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-fns.R
\name{mutation}
\alias{mutation}
\title{Mutation}
\usage{
mutation(parents, mutation_rate = 0.05, variation_parameter = 0.01)
}
\arguments{
\item{parents}{List of parents to apply the mutation step.}

\item{mutation_rate}{Mutation rate. Numeric value between 0-1 representing the probability of mutation of a single gene.}

\item{variation_parameter}{Random variation applied to every gene. Numeric value between 0-1.}
}
\description{
Applies the mutation step in the evolutionary algorithm
}
\details{
A colour is converted to RGB and each value is randomly drawn from a beta distribution with mean of the existing
red, green or blue values on the 0-1 scale and approximate standard deviation of \code{variation_parameter}. In short smaller
values are less variable and larger values more variable.

The \code{mutation_rate} is the probability any given colour will be replaced with a completely random one.
}
\examples{
pal <- random_palette(5, 2)
original <- map(pal, ~show_palette(.x, "original"))
mutated <- pal \%>\%
   mutation() \%>\%
   map(~show_palette(.x, "mutated"))
c(original, mutated) \%>\% wrap_plots
}
