% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-fns.R
\name{mutation}
\alias{mutation}
\title{Mutation}
\usage{
mutation(parents, mutation_rate = 0.05, variation_parameter = 0.01)
}
\arguments{
\item{parents}{List of parents to apply the mutation step.}

\item{mutation_rate}{Mutation rate. Numeric value between 0-1 representing the probability of mutation of a single colour.}

\item{variation_parameter}{Numeric value between 0-1. Random variation applied to every gene.}
}
\description{
Applies the mutation step in the evolutionary algorithm
}
\examples{
\dontrun{
pal <- random_palette(5, 2)
original <- map(pal, ~show_palette(.x, "original"))
mutated <- pal \%>\%
    mutation() \%>\%
    map(~show_palette(.x, "mutated"))
c(original, mutated) \%>\%
    wrap_plots
}
}
