% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-fns.R
\name{plot_palette}
\alias{plot_palette}
\title{Test plot}
\usage{
plot_palette(pal, aes = "fill")
}
\arguments{
\item{pal}{}
}
\description{
Plots a classic histogram using MPG data to test the colour palette
}
