% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-fns.R
\name{plot_palette}
\alias{plot_palette}
\title{Test plot}
\usage{
plot_palette(pal, aesthetic = "fill")
}
\arguments{
\item{pal}{Palette list or character vector}

\item{aesthetic}{'fill' or 'colour' aesthetic}
}
\description{
Plots MPG data displaying examples of the fill and colour aesthetics.
}
\examples{
\dontrun{
random_palette(5, 1) \%>\%
    plot_palette()
}
}
