% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-fns.R
\name{show_palette}
\alias{show_palette}
\title{Show palette}
\usage{
show_palette(pal, title = NULL, n = NULL, labels = FALSE,
  n_continuous = 3)
}
\arguments{
\item{pal}{Palette. Character vector.}

\item{title}{To add a title to the plot. Default NULL.}

\item{n}{Number of colours to show. Defaults to c(length(pal), 200)}

\item{labels}{Adds number labels to the colour palette.}

\item{n_continuous}{Number of colours from the palette the continuous gradient uses. Default 3.}
}
\description{
plots the discrete and continuous colour palettes
}
\details{
By default the continuous scale is set by taking 3 equally spaced colours along the colour palette. Can to set to more
or less if desired.
}
\examples{
random_palette(5, 2) \%>\%
    map(~show_palette(.x)) \%>\%
    wrap_plots()
}
