% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-fns.R
\name{show_palette}
\alias{show_palette}
\title{Show palette}
\usage{
show_palette(pal, n = NULL)
}
\arguments{
\item{n}{Number of colours to show. Defaults to c(length(pal), 200)}

\item{p}{Palette object}
}
\description{
Shows the palette
}
